requirements.txt
albumentations>=1.3
datasets>=2.14
diffusers>=0.24
einops>=0.6
ftfy>=6.1
kornia>=0.6
natsort>=8.4
numpy>=1.23
omegaconf>=2.3
opencv-python>=4.8
packaging>=23.0
pandas>=1.5
Pillow>=9.5
pytorch-fid>=0.3
pytorch-lightning==1.5.10
regex>=2023.0
requests>=2.31
scann>=1.2.10
streamlit>=1.28
taming-transformers-rom1504>=0.0.6
torch>=1.13
torch-geometric>=2.3
torchvision>=0.14
tqdm>=4.65
webdataset>=0.2
git+https://github.com/openai/CLIP.git
git+https://github.com/CompVis/latent-diffusion.git
